Opendata/requirements.txt
asn1crypto
chardet
django-rest-framework
django-tinymce
django-wysiwyg
gunicorn
pip-chill
pycparser
python-decouple
xlrd
